v3-arb-bot/requirements.txt
ccxt==1.120.34
python-dotenv==1.0.0
streamlit==1.26.0
pandas==2.1.0
pytest==7.4.0
flake8==6.0.0
requests==2.31.0
psycopg2-binary==2.9.10